meson: work around broken system PCRE2 dependency in macOS
macOS provides a PCRE2 library in base that is not usable and not
configured properly, as it installs a pkgconf module that
points to a non-existent pcre2.h header in /usr/local/include.

Detect that case and allow a fallback to a wrapped submodule
if the feature is enabled and that is possible, or print a
warning and disable the feature if the feature was set to "auto",
which is the new default.

To keep consistency with the cmake build system used on Windows,
add a special rule to re-enable the pcre2 feature there.

Helped-by: Eric Sunshine <sunshine@sunshineco.com>
Suggested-by: Eli Schwartz <eschwartz@gentoo.org>
Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>