ref-iterator-seek: correctly initialize the prefix_state for a new level
When cache_ref_iterator_seek() "jumps" to a middle of the sorted ref
list, it forgets to set the .prefix_state member of the new
(i.e. deeper) level it just initialized.  This later causes
cache_ref_iterator_advance() to look at this uninitialized member
to base its decision on what to do next.

Kyle Lippincott [*] and Jeff King noticed this with MSAN and
Valgrind, and Karthik Nayak as the original author located exactly
where the missing initialization is.

[*] <CAO_smVg9TDakUnubepjPGmLyOzW6n8Z=MDbnZKvkwN2=kN2RRw@mail.gmail.com>

Signed-off-by: Junio C Hamano <gitster@pobox.com>