diff: simplify --no-index taking advantage of the fact that prefix==NULL
Since the caller of the diff_no_index() function now undoes the
effect of setup.c:setup_git_directory_gently(), i.e. figuring out
where in the working tree the cwd of the process is, recording the
cwd relative to that top directory in the prefix, and chdir()ing up
to the top directory, we are now guaranteed to always have NULL in
the prefix.  When started outside a repository, prefix would already
be NULL.

Take advantage of this fact to simplify the implementation.

 - The code that conditionally calls prefix_filename(), except for
   "-" given from the command line, can go away, as there is nothing
   to prefix.

 - The temporary variable that points at the memory allocated by
   prefix_filename() for later freeing can go; so can the loop to
   free them.

 - As the prefix_filename() loop that had to special case "-" is
   gone, we do not rewrite "-" to another constant address that has
   the same string, file_from_standard_input[].  Teach get_mode()
   function to strcmp() between the path(s) and "-" directly for
   simplicity.

Signed-off-by: Junio C Hamano <gitster@pobox.com>